Fixed some issues detected by clang static analyzer.  See accompanying r943 commit on Easel, which has more information, including installation and usage of the clang static analyzer "scan-build". Almost all of these fixes are harmless allocation type mismatches.
hmmdmstr.c lines 1124/1125 were allocating the wrong size altogether,
and could potentially have caused memory overwrite when the daemon
gets a SHUTDOWN signal.

hmmdmstr.c line 952 has a very wrong looking allocation that may even
be a bug. But Farrar's code is fiddling around with pointers, so maybe
he was doing something deliberate. Left a note there.

p7_checkptmx.c line 98 bothers the analyzer but the code is correct;
analyzer can't see that we're dealing with hand-aligned SIMD __m128
vectors.

p7_scoredata.c line 73 appeared to be an error; MSV scores are uints,
not floats. But Travis's code after that allocation is also silently
(and bravely) casting uints to floats, I think, so there may be more
issues here.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4778 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ src/base/p7_scoredata.c
@@ -70,7 +70,7 @@ scoredata_GetMSVScoreArrays(P7_OPROFILE *om, P7_SCOREDATA *data, int do_opt_ext 
   p7_oprofile_GetMSVEmissionScoreArray(om, data->msv_scores);
 
   if (do_opt_ext) {
-    ESL_ALLOC(max_scores, (om->M + 1) * sizeof(float));
+    ESL_ALLOC(max_scores, (om->M + 1) * sizeof(uint8_t));
     for (i = 1; i <= om->M; i++) {
       max_scores[i] = 0;
       for (j=0; j<K; j++)
